Split views into business logic layer, api layer
